requirements.txt
torch
torchvision
inference-sdk
tqdm
pillow
pandas
opencv-python
matplotlib
requests
editdistance
scikit-learn